refactor dataset module
split into separate submodules under kraken.lib.dataset
